backend/requirements.txt
Flask
Flask-CORS
python-dotenv
google-generativeai